relocate: theme page to add reusable widgets folder to core
